% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateAndCreateMetaLocal.R
\name{createH5}
\alias{createH5}
\title{Creates metafiles for HDF5-files}
\usage{
createH5(data, file, dataset_name)
}
\arguments{
\item{data, }{contains metadata}

\item{file, }{contains file name}

\item{dataset_name, }{contains dataset name}
}
\description{
Creates metafiles for HDF5-files
}
