% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateAndCreateMetaLocal.R
\name{createIndexH5}
\alias{createIndexH5}
\title{Writes indexes to the file}
\usage{
createIndexH5(data, file)
}
\arguments{
\item{data, }{contains metadata}

\item{file}{contains the file name}
}
\value{
Returns NULL
}
\description{
Writes indexes to the file
}
\keyword{internal}
