% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateAndCreateMetaRemote.R
\name{createIndexH5Remote}
\alias{createIndexH5Remote}
\title{Creates HDF5-file with indexes}
\usage{
createIndexH5Remote(url)
}
\arguments{
\item{url, }{contains url to the root of counts files}
}
\value{
Returns NULL
}
\description{
Creates HDF5-file with indexes
}
