% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateAndCreateMetaLocal.R
\name{createMetaH5}
\alias{createMetaH5}
\title{Converts collection meta.txt files to meta.h5, putting them to the respective
collection folders}
\usage{
createMetaH5(counts_dir)
}
\arguments{
\item{counts_dir, }{contains directory name}
}
\value{
Returns NULL
}
\description{
Converts collection meta.txt files to meta.h5, putting them to the respective
collection folders
}
