% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateAndCreateMetaLocal.R
\name{createPriorityH5}
\alias{createPriorityH5}
\title{Creates HDF5-File with priority}
\usage{
createPriorityH5(counts_dir, force = FALSE, verbose = FALSE)
}
\arguments{
\item{counts_dir, }{contains counts directory}

\item{force}{logical value which lets function replace existing priority file}

\item{verbose}{logical value which determines a content of  the output.}
}
\value{
Returns NULL
}
\description{
Creates HDF5-File with priority
}
\keyword{internal}
