% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateAndCreateMetaLocal.R
\name{getCountsMetaPart}
\alias{getCountsMetaPart}
\title{Gets list  with metadata}
\usage{
getCountsMetaPart(counts_dir, collection_name, verbose)
}
\arguments{
\item{counts_dir, }{contains counts directory}

\item{collection_name}{contains name of the collection}

\item{verbose}{logical value which determines a content of  the output.}
}
\value{
list with metadata
}
\description{
Gets list  with metadata
}
