% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHSDSFileList.R
\name{getHSDSCollectionFileList}
\alias{getHSDSCollectionFileList}
\title{Returns list of HDF5-files of a collection}
\usage{
getHSDSCollectionFileList(
  url = "https://ctlab.itmo.ru/hsds/?domain=/counts",
  collection
)
}
\arguments{
\item{url, }{containing url of the server and root domain.}

\item{HDF5}{collection name}
}
\value{
List of HDF5-files of the collection
}
\description{
Returns list of HDF5-files of a collection
}
