% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHSDSFileList.R
\name{getHSDSCollectionsList}
\alias{getHSDSCollectionsList}
\title{Returns list of collections on HSDS-server}
\usage{
getHSDSCollectionsList(url = "https://ctlab.itmo.ru/hsds/?domain=/counts")
}
\arguments{
\item{url, }{containing url of the server and root domain.}
}
\value{
List of collections on the server
}
\description{
Returns list of collections on HSDS-server
}
