% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHSDSFileList.R
\name{getHSDSFileList}
\alias{getHSDSFileList}
\title{Returns list of all HDF5-files on HSDS-server}
\usage{
getHSDSFileList(
  url = "https://alserglab.wustl.edu/hsds/?domain=/counts",
  directory = NULL
)
}
\arguments{
\item{url, }{containing url of the server and root domain.}

\item{directory, }{containing name of the directory}
}
\value{
List of all HDF5-files on the server or all files of the collection
}
\description{
Returns list of all HDF5-files on HSDS-server
}
\examples{
url <- 'https://alserglab.wustl.edu/hsds/?domain=/counts'
getHSDSFileList(url)

}
