% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateAndCreateMetaRemote.R
\name{getIndexRemote}
\alias{getIndexRemote}
\title{Creates a data table with indexes and chunks of samples in rempte HDF5-files}
\usage{
getIndexRemote(src, collections)
}
\arguments{
\item{src, }{contains url of the server}

\item{collections, }{contains names of the collections}
}
\description{
Creates a data table with indexes and chunks of samples in rempte HDF5-files
}
