% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCountsFromH5file.R
\name{gsmToChunk}
\alias{gsmToChunk}
\title{Gets chunk from GSE identifiers.}
\usage{
gsmToChunk(samples)
}
\arguments{
\item{samples, }{containing a list of samples}
}
\value{
list of chunks
}
\description{
Gets chunk from GSE identifiers.
}
