% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCondition.R
\name{inferCondition}
\alias{inferCondition}
\title{Adds condition to the annotation.}
\usage{
inferCondition(es)
}
\arguments{
\item{es, }{contains ExpressionSet object}
}
\value{
Annotated ExpressionSet with conditions and replicates
}
\description{
Adds condition to the annotation.
}
\examples{
es <- inferCondition(es)

}
