% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCondition.R
\name{inferConditionImpl}
\alias{inferConditionImpl}
\title{Creates condition from the samples titles}
\usage{
inferConditionImpl(gse_titles)
}
\arguments{
\item{gse_titles, }{contains titles}
}
\value{
List of conditions and replicates
}
\description{
Creates condition from the samples titles
}
\keyword{internal}
