% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gct.R
\name{read.gct}
\alias{read.gct}
\title{Reads ExpressionSet from a GCT file.}
\usage{
read.gct(gct)
}
\arguments{
\item{gct}{Path to gct file}
}
\value{
ExpressionSet object
}
\description{
Only versions 1.2 and 1.3 are supported.
}
\examples{
es <- read.gct(system.file("testdata/gct/test.gct", package="phantasusLite"))
}
