% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gct.R
\name{readGct}
\alias{readGct}
\title{Reads ExpressionSet from a GCT file.}
\usage{
readGct(gct)
}
\arguments{
\item{gct}{Path to gct file}
}
\value{
ExpressionSet object
}
\description{
Only versions 1.2 and 1.3 are supported.
}
\examples{
es <- readGct(system.file("extdata/testdata/gct/test.gct", package="phantasusLite"))
}
