% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCondition.R
\name{removeRepeatWords}
\alias{removeRepeatWords}
\title{Removes repeated words from conditions}
\usage{
removeRepeatWords(titles)
}
\arguments{
\item{titles, }{contains titles}
}
\value{
titles without repeated words
}
\description{
Removes repeated words from conditions
}
