% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateAndCreateMetaLocal.R
\name{updateARCHS4meta}
\alias{updateARCHS4meta}
\title{Creates \code{meta.txt} file, which describes typical archs4 and archs4Zoo files.}
\usage{
updateARCHS4meta(
  archDir = file.path(getOption("phantasusCacheDir"), "counts/archs4")
)
}
\arguments{
\item{archDir}{path to directory with arch4 .h5 files.}
}
\value{
Returns NULL
}
\description{
Creates \code{meta.txt} file, which describes typical archs4 and archs4Zoo files.
}
\keyword{internal}
