% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateAndCreateMetaLocal.R
\name{updateDEE2meta}
\alias{updateDEE2meta}
\title{Creates \code{meta.txt} file, which describes typical dee2 files.}
\usage{
updateDEE2meta(
  destDir = file.path(getOption("phantasusCacheDir"), "counts/dee2")
)
}
\arguments{
\item{destDir}{path to directory with DEE2 .h5 files.}
}
\value{
Returns NULL
}
\description{
Creates \code{meta.txt} file, which describes typical dee2 files.
}
