% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateAndCreateMetaLocal.R
\name{updateIndexH5}
\alias{updateIndexH5}
\title{Updates indexes from HDF5-files}
\usage{
updateIndexH5(counts_dir, force = FALSE, verbose = FALSE)
}
\arguments{
\item{counts_dir, }{contains counts directory}

\item{force}{logical value which lets function replace existing index file}

\item{verbose}{logical value which determines a content of  the output.}
}
\value{
Returns NULL
}
\description{
Updates indexes from HDF5-files
}
\keyword{internal}
