% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateAndCreateMetaLocal.R
\name{validateCountsCollection}
\alias{validateCountsCollection}
\title{Validates counts collection}
\usage{
validateCountsCollection(collectionDir, verbose = FALSE)
}
\arguments{
\item{collectionDir}{contains directory name}

\item{verbose}{logical value which determines a content of  the output.}
}
\value{
false if collection is not valid
}
\description{
Validates counts collection
}
\keyword{internal}
