% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHSDSFileList.R
\name{getHSDSFileList}
\alias{getHSDSFileList}
\title{Returns list of all HDF5-files on HSDS-server}
\usage{
getHSDSFileList(url = "https://ctlab.itmo.ru/hsds/?domain=/counts")
}
\arguments{
\item{url, }{containing url of the server and root domain.}
}
\value{
List of all HDF5-files on the server
}
\description{
Returns list of all HDF5-files on HSDS-server
}
