% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCondition.R
\name{inferConditionImpl}
\alias{inferConditionImpl}
\title{Create condition from the samples title}
\usage{
inferConditionImpl(gse_titles)
}
\arguments{
\item{gse_titles, }{contains titles}
}
\description{
Create condition from the samples title
}
