% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCountsFromH5file.R
\name{loadCountsFromH5FileHSDS}
\alias{loadCountsFromH5FileHSDS}
\title{Load count matrix from remote HDF5-file}
\usage{
loadCountsFromH5FileHSDS(
  es,
  url,
  file,
  sample_id = NULL,
  gene_id = NULL,
  gene_id_type = NULL,
  sampleIndexes = NULL
)
}
\arguments{
\item{es, }{containing ExpressionSet loaded from GEO. Contains empty expression matrix.}

\item{url, }{containing url of the server and root domain.}

\item{file, }{containing name of the file}

\item{sample_id, }{containing path to the dataset with sample ids}

\item{gene_id, }{containing path to the dataset with gene ids}

\item{gene_id_type, }{containing path to the dataset with gene id type}

\item{sampleIndexes, }{containing sample indexes list}
}
\value{
ExpressionSet object with loaded count matrix
}
\description{
Load count matrix from remote HDF5-file
}
