% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCountsFromH5file.R
\name{loadCountsFromHSDS}
\alias{loadCountsFromHSDS}
\title{Load count matrix from HDF5-files.}
\usage{
loadCountsFromHSDS(es, url = "https://ctlab.itmo.ru/hsds/?domain=/counts")
}
\arguments{
\item{es, }{containing ExpressionSet loaded from GEO. Contains empty expression matrix.}

\item{url, }{containing url of the server and root domain.}
}
\description{
Load count matrix from HDF5-files.
}
